Seperate tests and increase coverage for email verification controller
